refactor: forgot-password.jsx renamed to *.tsx
